+++ src/selectors.js
@@ -1,21 +1,11 @@
-const rtlcss = require( 'rtlcss' )
+const isSelectorHasDir = ( selector = '' ) =>
+    !!selector.match( /^html\[dir(=".+")?\]/ )
 
 const isHtmlSelector = ( selector = '' ) =>
     !!selector.match( /^html/ )
 
 const isRootSelector = ( selector = '' ) =>
     !!selector.match( /:root/ )
-
-const getDirRule = ( rule, dir ) => {
-    const next = rule.next()
-    const selector = addDirToSelectors( rule.selector, dir )
-
-    if ( next && next.selector === selector ) {
-        return next
-    } else {
-        return rule.cloneAfter( { selector } ).removeAll()
-    }
-}
 
 const addDirToSelectors = ( selectors = '', dir ) => {
     let prefix
@@ -46,28 +36,9 @@ const addDirToSelectors = ( selectors = '', dir ) => {
     return selectors
 }
 
-const setRuleDir = ( rule, dir )=>
-    rule.selector = addDirToSelectors( rule.selector, dir )
-
-const rtlifyDecl = decl => {
-    const rtlResult = rtlcss.process( decl, null, null )
-    if ( rtlResult === decl.toString() ) return false
-
-    let [ prop, value ] = rtlResult.split( /:\s*/ )
-    return { prop, value }
-}
-
-const rtlifyRule = rule => {
-    const rtlResult = rtlcss.process( rule, null, null )
-
-    return ( rtlResult !== rule.toString() ) ? rtlResult : false
-}
-
 module.exports = {
+    isSelectorHasDir,
     isHtmlSelector,
     isRootSelector,
-    getDirRule,
-    setRuleDir,
-    rtlifyDecl,
-    rtlifyRule
+    addDirToSelectors
 }