+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.3.1] - 2016-10-31
+### Added
+- Fixed `@keyframes`-animations processing
 
 ## [0.3.0] - 2016-10-29
 ### Added
-- Process and flip keyframe-animations
+- Process and flip `@keyframes`-animations
 
 ## [0.2.0] - 2016-10-29
 ### Added